web-interface/requirements.txt
Flask>=3.0
paho-mqtt>=1.6
opencv-python-headless>=4.9
numpy>=1.26
pymongo>=4.6
